pyproject.toml
[project]
name = "basic-components"
version = "0.1.0"
description = "Re-usable server-side components based on shadcn/ui with JinjaX, Alpine.js and Tailwind CSS"
authors = [
    { name = "phernandez", email = "phernandez@basicmachines.co" },
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi[standard]>=0.115.4",
    "jinjax[whitenoise]>=0.47",
    "pyright>=1.1.385",
    "pytest-playwright>=0.5.2",
    "pytest>=8.3.3",
    "starlette-wtf>=0.4.5",
    "uvicorn>=0.32.0",
    "wtforms>=3.2.1",
    "jinja2>=3.1.3",
    "python-frontmatter>=1.1.0", # For markdown metadata
    "pygments>=2.17.2", # For code highlighting
    "loguru>=0.7.2",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "arel>=0.3.0",
    "websockets>=13.1",
    "markdown>=3.7",
    "markupsafe>=3.0.2",
    "watchfiles>=0.24.0",
    "icecream>=2.1.3",
    "pymdown-extensions>=10.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["basic_components"]

[tool.hatch.build]
include = [
    "basic_components/**/*.py",
    "components/**/*",
    "docs/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"